requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
pydantic==2.8.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
jinja2==3.1.4
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
httpx==0.27.2
pytest==8.3.2
bcrypt==3.2.2
email-validator
